+++ qupath-core/src/main/java/qupath/lib/objects/PathObject.java
@@ -23,9 +23,11 @@
 
 package qupath.lib.objects;
 
+import java.util.Objects;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import qupath.lib.common.ColorTools;
+import qupath.lib.common.LogTools;
 import qupath.lib.interfaces.MinimalMetadataStore;
 import qupath.lib.io.PathIO;
 import qupath.lib.measurements.MeasurementList;
@@ -383,7 +385,7 @@ public abstract class PathObject implements Externalizable, MinimalMetadataStore
 	 * @param list
 	 * @param toRemove
 	 */
-	private static <T> void removeAllQuickly(Collection<T> list, Collection<T> toRemove) {
+	private static <T> void removeAllQuickly(Collection<T> list, Collection<? extends T> toRemove) {
 		int size = 10;
 		if (!(toRemove instanceof Set)  && toRemove.size() > size) {
 			toRemove = new HashSet<>(toRemove);
@@ -403,10 +405,12 @@ public abstract class PathObject implements Externalizable, MinimalMetadataStore
 
 	/**
 	 * Remove a single object from the child list of this object.
-	 * @param pathObject
+	 * @param pathObject the object to remove
 	 * @since v0.4.0
+	 * @throws NullPointerException if the object to remove is null
 	 */
-	public void removeChildObject(PathObject pathObject) {
+	public void removeChildObject(PathObject pathObject) throws NullPointerException {
+		Objects.requireNonNull(pathObject, "PathObject passed to removeChildObject(PathObject) must not be null!");
 		if (!hasChildObjects())
 			return;
 		if (pathObject.parent == this)
@@ -417,10 +421,12 @@ public abstract class PathObject implements Externalizable, MinimalMetadataStore
 	
 	/**
 	 * Remove multiple objects from the child list of this object.
-	 * @param pathObjects
+	 * @param pathObjects the objects to remove
 	 * @since v0.4.0
+	 * @throws NullPointerException if the collection passed as a parameter is null
 	 */
-	public synchronized void removeChildObjects(Collection<PathObject> pathObjects) {
+	public synchronized void removeChildObjects(Collection<? extends PathObject> pathObjects) throws NullPointerException {
+		Objects.requireNonNull(pathObjects, "Collection passed to removeChildObjects(Collection) must not be null! Maybe you want to call removeAllChildObjects() instead?");
 		if (!hasChildObjects())
 			return;
 		for (PathObject pathObject : pathObjects) {
@@ -432,12 +438,11 @@ public abstract class PathObject implements Externalizable, MinimalMetadataStore
 		}
 	}
 
-	
 	/**
 	 * Remove all child objects.
-	 * @since v0.4.0
+	 * @since v0.8.0
 	 */
-	public void clearChildObjects() {
+	public void removeAllChildObjects() {
 		if (!hasChildObjects())
 			return;
 		synchronized (childList) {
@@ -447,6 +452,18 @@ public abstract class PathObject implements Externalizable, MinimalMetadataStore
 			}
 			childList.clear();
 		}
+	}
+
+
+	/**
+	 * Remove all child objects.
+	 * @since v0.4.0
+	 * @deprecated use instead {@link #removeAllChildObjects()}
+	 */
+	@Deprecated(since="0.8.0")
+	public void clearChildObjects() {
+		LogTools.warnOnce(logger, "clearChildObjects() has been replaced by removeAllChildObjects()");
+		removeAllChildObjects();
 	}
 
 	
@@ -989,8 +1006,6 @@ public abstract class PathObject implements Externalizable, MinimalMetadataStore
 				cachedUnmodifiableChildren = Collections.unmodifiableCollection(childList);
 			}
 		}
-//			childList = new TreeSet<PathObject>(DefaultPathObjectComparator.getInstance());
-//			childList = new ArrayList<PathObject>();
 	}
 	
 	/**